create channel_splitter.rs
